refactor: adapt to review
