+++ src/Runtimes/Runtimes.php
@@ -60,110 +60,3 @@ class Runtimes
         return $withVersion;
     }
 }
-
-class Runtime
-{
-    /**
-     * @var string
-     */
-    protected $key;
-
-    /**
-     * @var string
-     */
-    protected $name;
-
-    /**
-     * @var Version[]
-     */
-    protected $versions = [];
-
-    public function __construct(string $key, string $name)
-    {
-        $this->key = $key;
-        $this->name = $name;
-        $this->versions;
-    }
-
-    /**
-     * Adds new version to runtime.
-     * 
-     * @param string $version
-     * @param string $base
-     * @param string $image
-     * @param string[] $supports
-     */
-    public function addVersion(string $version, string $base, string $image, array $supports): void
-    {
-        $this->versions[] = new Version($version, $base, $image, $supports);
-    }
-
-    /**
-     * List runtime with all parsed Versions.
-     * 
-     * @return array[]
-     */
-    public function list(): array
-    {
-        $list = [];
-        foreach ($this->versions as $version) {
-            $key = "{$this->key}-{$version->version}";
-            $list[$key] = array_merge(
-                [
-                    'name' => $this->name,
-                    'logo' => "{$this->key}.png",
-                ],
-                $version->get()
-            );
-        }
-
-        return $list;
-    }
-}
-
-class Version
-{
-    /**
-     * @var string
-     */
-    public $version;
-
-    /**
-     * @var string
-     */
-    public $base;
-
-    /**
-     * @var string
-     */
-    public $image;
-
-    /**
-     * @var array
-     */
-    public $supports;
-
-    public function __construct(string $version, string $base, string $image, array $supports)
-    {
-        $this->version = $version;
-        $this->base = $base;
-        $this->image = $image;
-        $this->supports = $supports;
-    }
-
-    /**
-     * Get parsed Version.
-     * 
-     * @return (array|string)[]
-     */
-    public function get(): array
-    {
-        return
-            [
-                'version' => $this->version,
-                'base' => $this->base,
-                'image' => $this->image,
-                'supports' => $this->supports
-            ];
-    }
-}
